+++ private/Get-AzDFV2Credential.ps1
@@ -5,13 +5,6 @@ function Get-AzDFV2Credential {
     )
     Write-Debug "BEGIN: Get-AzDFV2Credential"
 
-    # Retrieve all credentials via API without parsing
-    $token = Get-AzAccessToken -ResourceUrl 'https://management.azure.com'
-    $tokenStr = [System.Net.NetworkCredential]::new('', $token.Token).Password
-    $authHeader = @{
-        'Content-Type'  = 'application/json'
-        'Authorization' = 'Bearer ' + $tokenStr
-    }
     $url = "$($script:BaseApiUrl)$($adfi.DataFactoryId)/credentials?api-version=2018-06-01"
 
     # Retrieve all credentials via Rest API